+++ CHANGELOG.md
@@ -22,6 +22,7 @@ No 32 bit builds are now available anymore. Only 64 bit (Intel and ARM) are supp
 - Add typings for external modules, remove the internal custom ones for `bcp-47` and `nspell`.
 - Reenabled the Pandoc and XeLaTeX options in preferences.
 - Removed the generic IPC call from the config provider and replaced it with consistent checks as to which options have actually changed. This increases the performance after configuration saves.
+- Migrate the config provider to TypeScript.
 
 # 1.8.7